refactor[#137] : NavigationBar 경로 수정
