+++ src/pages/sign-up/user-info/page.tsx
@@ -1,5 +1,5 @@
-import { BackwardNavigationBar } from "@/widgets/navigationbar/ui";
 import { UserInfoRegistrationForm } from "@/features/auth/ui";
+import { BackwardNavigationBar } from "@/shared/ui/navigationbar";
 
 const UserInfoRegistrationPage = () => {
   return (
